+++ collective/elasticsearch/indexes.py
@@ -162,7 +162,7 @@ class EDateIndex(BaseIndex):
             return {'range': {name: {'gte': first}}}
         elif range_ == 'max':
             return {'range': {name: {'lte': first}}}
-        elif range_ == 'minmax' and type(query) in (list, tuple) and \
+        elif range_ in ('min:max', 'minmax') and type(query) in (list, tuple) and \
                 len(query) == 2:
             return {'range':
                      {name: {